Revert "refactor: move readiness implementation to @percy/sdk-utils (PER-7348)"
This reverts commit d43c9bae0ca3a689b28a0a24f4bc6ed6e7eccaa1.
