+++ packages/sdk-utils/package.json
@@ -16,7 +16,6 @@
   },
   "files": [
     "dist",
-    "src/readiness-browser.js",
     "test/server.js",
     "test/client.js",
     "test/helpers.js"
@@ -28,7 +27,6 @@
       "node": "./dist/index.js",
       "default": "./dist/bundle.js"
     },
-    "./readiness-browser": "./src/readiness-browser.js",
     "./test/helpers": {
       "node": "./test/helpers.js",
       "default": "./test/client.js"
